feat: add story form and settings
